attempting cufftMakePlanMany64 pass 4: changed batch to long long int

+++ include/transforms/ffter.hpp
@@ -53,7 +53,7 @@ public:
 
 class CuFFTerR2C: public CuFFTer {
 public:
-  CuFFTerR2C(unsigned int size, unsigned int batch=1)
+  CuFFTerR2C(unsigned int size, long long int batch=1)
     :CuFFTer()
   {
     this->size = size;
@@ -62,7 +62,9 @@ public:
     cufftResult error1 =  cufftCreate(&fft_plan);
     ErrorChecker::check_cufft_error(error1);
 
-    cufftResult error2 = cufftMakePlanMany64(fft_plan, static_cast<long long>(1), n, NULL, static_cast<long long>(1), size, NULL, static_cast<long long>(1), size, CUFFT_R2C, static_cast<long long int>(batch), workSize);
+    cufftResult error2 = cufftMakePlanMany64(fft_plan, static_cast<long long>(1), n, 
+      NULL, static_cast<long long>(1), size, NULL, static_cast<long long>(1), size, 
+      CUFFT_R2C, static_cast<long long int>(batch), workSize);
     //cufftResult error = cufftPlan1d(&fft_plan, size, CUFFT_R2C, batch);
     ErrorChecker::check_cufft_error(error2);
   }
@@ -76,7 +78,7 @@ public:
 
 class CuFFTerC2R: public CuFFTer {
 public:
-  CuFFTerC2R(unsigned int size, unsigned int batch=1)
+  CuFFTerC2R(unsigned int size, long long int batch=1)
     :CuFFTer()
   {
     this->size = size;
@@ -86,7 +88,9 @@ public:
     cufftResult error1 =  cufftCreate(&fft_plan);
     ErrorChecker::check_cufft_error(error1);
 
-    cufftResult error2 = cufftMakePlanMany64(fft_plan, static_cast<long long>(1), n, NULL, static_cast<long long>(1), size, NULL, static_cast<long long>(1), size, CUFFT_C2R, static_cast<long long int>(batch), workSize);
+    cufftResult error2 = cufftMakePlanMany64(fft_plan, static_cast<long long>(1), n, 
+      NULL, static_cast<long long>(1), size, NULL, static_cast<long long>(1), size, 
+      CUFFT_C2R, static_cast<long long int>(batch), workSize);
 
     ErrorChecker::check_cufft_error(error2);
   }
